Fix `cargo deny` via a `cargo update`
There is a newer patch version of `unicode-ident`, though it incurs a newer
license I have yet to review should be allowed. If users which to use the
updated version (as our `Cargo.lock` is only for our internal tracking and
doesn't propagate to users, and we offer no binaries), they can simply run
`cargo update` on their end.

`cargo deny` did raise warnings on unused license allowances (which are okay,
they aren't harmful), but I commented them out so it's more obvious which
licenses are in fact in-tree.