`thiserror 2` (#12)
`thiserror 1` is still in tree due to the `dkg`/`modular-frost` crates, but
those are only included when the `multisig` feature is enabled.

Per https://github.com/monero-oxide/monero-oxide/issues/5, it may be better to
replace `cfg_attr(feature = "std", ...)` with a flag set by `rustversion`
(already in-tree as `thiserror` itself relies on it). I can redo this commit
if preferred at this time, else we can restore these removed `cfg_attr` later.