Add CI to check the MSRV (#18)
* Add CI to check the MSRV

* Attempt a check_msrv which is 'only' a single line

The GitHub CI doesn't like the prior version, potentially as it was a
multi-line function and the YML -> bash pipeline isn't ready for that? This
version represents as a single line accordingly.

* Simplify calling `check_msrv`

Also restore the declared `rust-version` to accelerate future calls to
`cargo msrv find`.

* Add echo statements to debug where the msrv action fails

* Use `ls` to clarify what directory the CI is in

* Restore multi-line function

Another workflow shows it SHOULD work without issue (thanks plowsof for the
reference).

* Automatically iterate over all members of the workspace

* Move the while loop into a function, add missing/correct `rust-version`
specifications

The while loop would continue iterating, even if a command failed. Now that
it's in a function, we have access to `return` to propagate any error codes.

The missing `rust-version` specifications were identified thanks to this
workflow being added.

* Inline the bash functions' comments since we're no longer attempting a
single-line variant