Misc nits from the past few days (#94)
* More aggressively `ZeroizeOnDrop`

The only one which is questionable should be `ZeroizeOnDrop` for
`SignableTransaction`. As it _already_ had `ZeroizeOnDrop`
implemented for most of its internal fields however, I don't really mind
extending to it all its odds and ends.

This was primarily due to a report received on Immunefi noting that
`ClsagMultisig` did not implement `ZeroizeOnDrop`, which was less aggresive
than our usual pattern.

* Fix reachable panic on maliciously crafted ScannableBlock

Received as a bug report on Immunefi (confirmed as a low severity).

* Consolidate documentation on the ring signatures' transcript

We elsewhere documented the message had preconditions yet didn't immediately
here yet.

* Use `saturating_add` instead of `wrapping_add`

This doesn't matter here as the values are a single byte, stored in a `usize`,
and we perform the check the value isn't too big sufficiently early. The safer
pattern is `saturating_add` however.

* Limit amount of arbitrary chunks as possible due to codifying policy

https://github.com/monero-oxide/monero-oxide/pull/77 limited the amount of
arbitrary data we acknowledge to only the data which fits under policy. As each
chunk takes at least one byte to declare (they should actually take at least
three due to the nonce tag, length, and arbitrary data tag), we can bound the
amount of chunks accordingly.

* Add note to `ExtraField::Nonce`

The `Extra::push_nonce` API (internal) already documented there's a length
bound being ignored. This is external with `ExtraField::Nonce`. While we don't
support pushing `ExtraField` onto `Extra`, one could still create an instance
of `ExtraField` which cannot be round-tripped (written and read). We don't
'fix' this at this time, solely note the peculiarity.

* Increase the amount of rounds allowed to sample decoys when testing

https://github.com/monero-oxide/monero-oxide/actions/runs/18291851745/job/52080775476
hit this, which is a first as this limit was supposed to be somewhat absurd.
Apparently, it's not so absurd it's unreachable after years of running the CI.

* Remove odd newline in docstring