Add `monero-epee` (#90)
* Import `monero-epee` as currently published on crates.io

Slight edits made to documentation and so it compiles. While this supported
`edition = "2021"`, `rust-version = "1.54"`, inclusion in the workspace sets a
MSRV of 1.60.

* Split `monero-epee` into three source files for clarity

Adds _extensive_ documentation to the stack and achieves a panic-free API.

* Add explicit definition of compatibility to `monero-epee`

* Distinguish header from version, add provenance to each

* Add provenance to `read_varint`, move low-level `read_*` functions to their own file

* Document provenance for `read_key`, error on empty keys

* Replace the `Seek` API for `serde_json::Value`-esque API

Triples the size of the decoder from ~900 bytes to ~2700 bytes but should be
practically usable, without risks prior incurred. Additionally, reuse of the
decoder's allocation is allowed when performing indexes.

* Misc documentation clarifications, inline where appropriate

* Replace `index` with `iterate` due to the O(n) cost of it

* Remove `field` for `fields`, with `FieldIterator`

* Add shorter lifetimes to children of parents

Them having the same lifetime meant only child could exist, as it had the same
lifetime as the parent it mutably borrowed.

* Correct bound on `self` in `fn next`

* Assign `self` its own lifetime entirely in `fn next`

* Remove unnecessary requirement that `'a` outlives `'b`

* Revert prior commits messing with lifetimes, remove snapshots, reversion

The prior commits were insufficient to resolve the issues when using.

Snapshots and reversion aren't worth the complexity. We currently only offer
iterators, exposing the O(n) cost of accesses to users. With those, which are
ordered, there's no reason for snapshots, which only existed to support
unordered indexes.

* Fix lifetimes, when advancement occurs

* Update claim in README about memory usage

* Clarify docstring on `Epee::fields`

* Be generic to how bytes are represented

* Have `single_step` return an `enum`

Avoids having to peek VarInt lengths before beginning decoding of an object.

Introduces `InternalError` to represent if `single_step` does not return the
expected result. While we could panic, as it should be obvious, it's safer to
use `InternalError`.

* Remove the `Clone` bound from `BytesLike`

It's just exceptionally prone to be misused. `Clone` could be implemented on a
type for which it's expensive to extremely negative effects. We can offer 99%
of our functionality _without_ requiring it.

* Add test demonstrating `Epee::fields` cannot be called twice

* Correct maintenance of the stack

* clippy

* Tweak documentation

* Update `monero-oxide-no-std-tests` to verify both `core` and `alloc` builds

Justified due to how `monero-epee` should work on `core` alone.

This also updates the version of `monero-epee` now that it should be ready for
merge, and has had significantly changed since it started.

* Tweak `Epee` to infallibly return an `EpeeEntry` instead of fallibly returning a `FieldsIterator`

This single layer of infallible indirection allows considering the root object
itself as an `EpeeEntry`, making the API more consistent.

* Use an explicit error for if `Epee::entry` is called multiple times

Avoids a wonky rule with lifetimes.

* Move definition of `FieldIterator` in `monero-epee`

* Fix MSRV for `monero-epee`

* Remove `BytesLike::len`

This allows supporting containers for which we don't know the length of ahead
of time.

* Update commentary for `<&[u8] as BytesLike>::read_into_slice`

* `cargo +nightly fmt`

* Pack `TypeOrEntry` into `u4`

Saves ~5% of the decoder's size.

* Use `usize` to demark the lengths of objects, arrays

Equivalent on 64-bit platforms but produces much better code on 32-bit
platforms.

* `AsRef<B> for String<'encoding, B>`

* Remove `AsRef`, `AsMut` for `String`

It was unsound as the `AsMut` would allow taking from the bytes underneath,
while we never updated the length stored in the `String`.

* Export `String`

* Fix determination of the shift within `PackedTypes::get`

* Update citation for the EPEE specification

@jeffro256's wonderful description was upstreamed into Monero proper.

Also cites a PR to remove the array type we don't support from Monero, which
would make our implementation complete.

* Further caveat the definition of compatibility

Hedges against `epee` supporting serializing to an arbitrary depth (if it only
applies the recursion depth limit when deserializing).

* `doc_auto_cfg` -> `doc_cfg`

* Respond to PR feedback and ensure `error` is respected in all `pub fn`

* Correct dirty lockfile